% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{add_context}
\alias{add_context}
\title{Add context inside of tests name}
\usage{
add_context(desc, ...)
}
\arguments{
\item{desc}{Basic test name or description of the feature.}

\item{...}{Variables that should be added as context of the test.}
}
\value{
An enriched description with variables values.
}
\description{
This is particularly useful when using expectations inside of a loop.
}
\examples{
vars <- list(x = 5, y = 2)
z <- 3
"addition gives only positive values" |> add_context(vars, my_z = z)
}
