% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_project.R
\name{build_switch_project}
\alias{build_switch_project}
\title{Create a map allowing to switch project position}
\usage{
build_switch_project(maps)
}
\arguments{
\item{maps}{Named list of filepath you want to be able to switch in.}
}
\value{
A named list of functions that can be called to switch to a project.
The names are the same as in \code{maps}.
}
\description{
This is a hacky way to make the \code{here} package think that we are in another
directory. Can be useful if you want to check snippets of code from multiple
projects, but it is not advised in the long term.
}
\examples{
\dontrun{
switch_proj <- build_switch_project(
  "path_a" = "C:/my_first_project/file_a.R",
  "path_b" = "C:/my_second_project/file_b.R"
)
# Switch to the first project
switch_proj$path_a()
# ...

# Switch to the second project
switch_proj$path_b()
}
}
