% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_correct_class}
\alias{check_correct_class}
\title{Check that the argument is from the expected class}
\usage{
check_correct_class(x, expected_class, example_constructors)
}
\arguments{
\item{x}{Argument we want to test if it is from the correct class.}

\item{expected_class}{Name of the expected class.}

\item{example_constructors}{Example functions allowing to construct the
correct class.}
}
\description{
Otherwise it returns an error.
}
\examples{
example_fn <- function(my_value) {
  check_correct_class(my_value, "character", "character")
  my_value < "abc"
}
example_fn("1")
example_fn(1)
}
