% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_match_one}
\alias{check_match_one}
\title{Check that only one element was given and belongs to a list of values}
\usage{
check_match_one(x, values)
}
\arguments{
\item{x}{Element to check.}

\item{values}{List of allowed values.}
}
\description{
Check that only one element was given and belongs to a list of values
}
\examples{
example_fn <- function(x) {
  check_match_one(x, c("a", "b", "c"))
  x
}
example_fn("a")
example_fn("d")
}
