% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_subset}
\alias{check_subset}
\title{Check that a set of value is a subset of another one}
\usage{
check_subset(x, values)
}
\arguments{
\item{x}{Element to check.}

\item{values}{List of allowed values.}
}
\description{
Check that a set of value is a subset of another one
}
\examples{
example_fn <- function(x) {
  check_subset(x, c("a", "b", "c"))
  x
}
example_fn(c("a", "b"))
example_fn(c("a", "d"))
}
