% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{create_factor}
\alias{create_factor}
\title{Create a factor from key value pairs}
\usage{
create_factor(...)
}
\arguments{
\item{...}{A list of name-value pairs.}
}
\value{
A factor
}
\description{
Note that if your name-value pairs were already stored as a vector \code{x}, you
could simply create it with \code{factor(x, levels = x)}. This is just some
syntactic sugar over this approach.
}
\examples{
create_factor(
  "a" = "A",
  "b" = "B"
)
}
