% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_table.R
\name{dt_tidy_statistics}
\alias{dt_tidy_statistics}
\title{Compute statistics by groups on a data.table, placed in tidy format}
\usage{
dt_tidy_statistics(dt, cols, groups, func_map)
}
\arguments{
\item{dt}{A \code{data.table} object}

\item{cols}{Column names to compute stats}

\item{groups}{Column names to group statistics}

\item{func_map}{A named list of statistical functions to apply, names of the list will
appear in the \code{statistics} column.}
}
\value{
A \code{data.table} object containing:
\itemize{
\item \code{{groups}} The groups.
\item \code{statistics} Name of the statistic.
\item \code{variable} Name of the variable.
\item \code{value} Name of the associated value.
}
}
\description{
Compute statistics by groups on a data.table, placed in tidy format
}
\examples{
dt_example <- data.table::data.table(
  x = c(1, 1, 1, 1, 2, 2, 2, 2),
  y = c(1, 2, 3, 4, 5, 6, 7, 8),
  g = c("a", "a", "a", "a", "b", "b", "b", "b"),
  w = c("c", "c", "d", "d", "c", "c", "d", "d")
)
dt_tidy_statistics(dt_example, c("x", "y"), c("g", "w"), list("mean" = mean))
}
