% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{factor_as_numeric}
\alias{factor_as_numeric}
\title{Convert factor object with a numerical name to its numeric value}
\usage{
factor_as_numeric(x)
}
\arguments{
\item{x}{A factor, whose levels correspond to \code{numeric} values.}
}
\value{
A vector of type \code{numeric}.
}
\description{
The naive approach of doing \code{as.numeric()} on a factor will not consider the
names of the levels. The purpose of this function is to use them instead of
the underlying numeric structure of the factor.
}
\examples{
factor_as_numeric(factor(c("3", "1", "10")))
}
