% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_signif.R
\name{fmt_signif}
\alias{fmt_signif}
\title{Format numeric values according to the number of significant digits}
\usage{
fmt_signif(x, digits)
}
\arguments{
\item{x}{A \code{numeric} vector.}

\item{digits}{Integer indicating the number of significant digits to be used.}
}
\value{
The formatted \code{numeric} vector as a \code{character} vector, where each
number is displayed with the right amount of significant digits.
}
\description{
The numbers will be rounded and converted to character. This allows to
display trailing zeros in the number.
}
\examples{
fmt_signif(21.35, 3)
fmt_signif(1.3, 3)
}
