% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{labs_dict}
\alias{labs_dict}
\title{Modify plot labels using a data dictionary}
\usage{
labs_dict(plot, dict)
}
\arguments{
\item{plot}{\code{ggplot} Object.}

\item{dict}{Data dictionary in the form of a named \code{character} vector.}
}
\value{
A relabeled \code{ggplot} object.
}
\description{
In general it is better to use human readable labels in plots, but changing
the name of the column variables
}
\examples{
dict <- c("my_x" = "X name", "my_y" = "Y name", "my_c" = "Attribute")
df_ex <- data.frame(
 my_x = c(1, 2, 3),
 my_y = c(4, 5, 6),
 my_c = c("a", "a", "b")
)
my_plot <- ggplot2::ggplot(
 df_ex,
 ggplot2::aes(x = my_x, y = my_y, color = my_c)
) +
ggplot2::geom_point()

labs_dict(my_plot, dict)
}
