% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_update.R
\name{left_join_update}
\alias{left_join_update}
\title{Left join with column update}
\usage{
left_join_update(x, y, by = NULL, ...)
}
\arguments{
\item{x}{First data frame.}

\item{y}{Second data frame.}

\item{by}{Vector of columns to join by.}

\item{...}{Additionnal parameters.}
}
\value{
A data frame with the left join and column update applied.
}
\description{
It applies a left join and update columns already existing inside the
first table by the values of columns in the second table.
}
