% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rubintools-package.R
\docType{package}
\name{rubintools-package}
\alias{rubintools}
\alias{rubintools-package}
\title{Personal Collection of Helper Functions}
\description{
This package is a collection of miscellaneous helper functions.
}
\author{
\strong{Maintainer}: Jean Rubin \email{jean.rubin@example.com}

}
