% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{warn_missing_arg}
\alias{warn_missing_arg}
\title{Warn about the use of default values in \code{arg_match()}}
\usage{
warn_missing_arg(x)
}
\arguments{
\item{x}{possible values for the missing argument. The default value
corresponds to the first one.}
}
\description{
Warn about the use of default values in \code{arg_match()}
}
\examples{
example_fn <- function(my_value = c("a", "b")) {
  if (missing(my_value)) { warn_missing_arg(my_value) }
  my_value <- rlang::arg_match(my_value)
  my_value
}
example_fn()
example_fn("a")
}
