% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winsorize.R
\name{winsorize}
\alias{winsorize}
\title{Winsorization of a vector of values}
\usage{
winsorize(x, q_min = 0.025, q_max = 0.975)
}
\arguments{
\item{x}{A vector of numerical values.}

\item{q_min}{Lower bound quantile.}

\item{q_max}{Upper bound quantile.}
}
\value{
A vector such that every value under or over the given quantile bound
is clamped to the bound.
}
\description{
Winsorization of a vector of values
}
\examples{
winsorize(1:10, q_min = 0.2, q_max = 0.8)
}
